deployment/requirements.txt
Flask==3.1.0
tensorflow==2.17.0
numpy==1.24.4
Pillow==10.2.0
gunicorn
google-cloud-storage